Set the color palette in index, delete colors.js
