--- colors.js
@@ -1,25 +1,0 @@
-// Color palette
-const Colors = {
-    // Main
-    BlackMain: '#333',
-    BlackSecond: '#222',
-    BlackThird: '#444',
-    Blue: '#0052CC',
-    Cyan: '#429eff',
-    GrayBackground: '#CCC',
-    GrayDisabled: '#aaa',
-    GrayHover: '#555',
-    White: '#FFF',
-
-    // Shadows
-    Shadow13: rgba(150, 150, 150, 0.13),
-    Shadow25: rgba(150, 150, 150, 0.25),
-
-    // Others
-    BackgroundSelected: '#353a44',
-    BackgroundSelectedHover: '#484f5d',
-    TextSelected: '#9494ff',
-    TextSubtle: '#d2d2d2',
-}
-
-export default Colors
