+++ index.js
@@ -1,4 +1,26 @@
-import Colors from "./colors"
+// Set a color palette for the dark theme
+const Colors = {
+    // Main
+    BlackMain: '#333',
+    BlackSecond: '#222',
+    BlackThird: '#444',
+    Blue: '#0052CC',
+    Cyan: '#429eff',
+    GrayBackground: '#CCC',
+    GrayDisabled: '#aaa',
+    GrayHover: '#555',
+    White: '#FFF',
+
+    // Shadows
+    Shadow13: rgba(150, 150, 150, 0.13),
+    Shadow25: rgba(150, 150, 150, 0.25),
+
+    // Others
+    BackgroundSelected: '#353a44',
+    BackgroundSelectedHover: '#484f5d',
+    TextSelected: '#9494ff',
+    TextSubtle: '#d2d2d2',
+}
 
 // Detect Jira website's header and add the new button
 let header = document.getElementsByTagName("header")[0].childNodes[1]
